requirements.txt
openai==0.27.2
pandas==1.3.4
matplotlib==3.4.3
langchain==0.0.152
python-dotenv==1.0.0
seaborn